+++ orchestration/src/main/java/com/sap/ai/sdk/orchestration/spring/OrchestrationChatModel.java
@@ -6,6 +6,7 @@ import static com.sap.ai.sdk.orchestration.model.MessageToolCall.TypeEnum.FUNCTI
 import com.sap.ai.sdk.orchestration.AssistantMessage;
 import com.sap.ai.sdk.orchestration.OrchestrationChatCompletionDelta;
 import com.sap.ai.sdk.orchestration.OrchestrationClient;
+import com.sap.ai.sdk.orchestration.OrchestrationModuleConfig;
 import com.sap.ai.sdk.orchestration.OrchestrationPrompt;
 import com.sap.ai.sdk.orchestration.SystemMessage;
 import com.sap.ai.sdk.orchestration.ToolMessage;
@@ -15,8 +16,9 @@ import com.sap.ai.sdk.orchestration.model.MessageToolCallFunction;
 import java.util.List;
 import java.util.Map;
 import java.util.function.Function;
-import java.util.stream.Collectors;
 import javax.annotation.Nonnull;
+import javax.annotation.Nullable;
+import lombok.Setter;
 import lombok.extern.slf4j.Slf4j;
 import lombok.val;
 import org.springframework.ai.chat.messages.AssistantMessage.ToolCall;
@@ -24,9 +26,8 @@ import org.springframework.ai.chat.messages.Message;
 import org.springframework.ai.chat.messages.ToolResponseMessage;
 import org.springframework.ai.chat.model.ChatModel;
 import org.springframework.ai.chat.model.ChatResponse;
+import org.springframework.ai.chat.prompt.ChatOptions;
 import org.springframework.ai.chat.prompt.Prompt;
-import org.springframework.ai.model.tool.DefaultToolCallingManager;
-import org.springframework.ai.model.tool.ToolCallingChatOptions;
 import reactor.core.publisher.Flux;
 
 /**
@@ -38,9 +39,7 @@ import reactor.core.publisher.Flux;
 public class OrchestrationChatModel implements ChatModel {
   @Nonnull private final OrchestrationClient client;
 
-  @Nonnull
-  private final DefaultToolCallingManager toolCallingManager =
-      DefaultToolCallingManager.builder().build();
+  @Setter @Nullable private OrchestrationChatOptions defaultOptions;
 
   /**
    * Default constructor.
@@ -63,29 +62,23 @@ public class OrchestrationChatModel implements ChatModel {
 
   @Nonnull
   @Override
+  public ChatOptions getOptions() {
+    if (defaultOptions != null) {
+      return defaultOptions;
+    }
+    return new OrchestrationChatOptions(new OrchestrationModuleConfig());
+  }
+
+  @Nonnull
+  @Override
   public ChatResponse call(@Nonnull final Prompt prompt) {
     if (prompt.getOptions() instanceof OrchestrationChatOptions options) {
 
       val orchestrationPrompt = toOrchestrationPrompt(prompt);
       val response =
           new OrchestrationSpringChatResponse(
-              client.chatCompletion(orchestrationPrompt, options.getConfig()));
+              client.chatCompletion(orchestrationPrompt, options.getConfigWithCallbacks()));
 
-      if (ToolCallingChatOptions.isInternalToolExecutionEnabled(prompt.getOptions())
-          && response.hasToolCalls()) {
-
-        if (log.isDebugEnabled()) {
-          val tools = response.getResult().getOutput().getToolCalls();
-          val toolsStr = tools.stream().map(ToolCall::name).collect(Collectors.joining(", "));
-          log.debug("Executing {} tool call(s) - {}.", tools.size(), toolsStr);
-        }
-
-        val toolExecutionResult = toolCallingManager.executeToolCalls(prompt, response);
-
-        // Send the tool execution result back to the model.
-        log.debug("Re-invoking LLM with tool execution results.");
-        return call(new Prompt(toolExecutionResult.conversationHistory(), prompt.getOptions()));
-      }
       return response;
     }
     throw new IllegalArgumentException(